+++ CHANGES.md
@@ -5,6 +5,7 @@ Changes in 2.9.5
 ----------------
 - Fix #152 Cookie on localhost/ (root path) and IE in Warehouse.php
 - Add jQuery Chosen 1.5.1 plugin in assets/js/jquery-chosen
+- Add ChosenSelectInput() function in Inputs.php
 
 Changes in 2.9.4
 ----------------